refactor(audit): extract snapshot+reflection out of controller into a service
Addresses Copilot's layering concern (PR #3591 thread): the System Audit
Events controller previously injected @PersistenceContext EntityManager
and ran reflection-based entity loading inline — persistence work in the
controller violates the project's "services own @Transactional" rule.

Move it to a new AuditEntitySnapshotService:
  - @Service / @Transactional(readOnly = true)
  - Owns REF_TABLE_TO_ENTITY_CLASS and SNAPSHOT_FIELDS_BY_REF_TABLE
  - Public surface: loadSnapshot(refTableName, refId) and
    loadFieldValues(refTableName, refId, fieldNames)
  - PII allow-list semantics preserved exactly (Patient ⊆ nationalId/
    externalId/gender; Person ⊆ firstName/lastName; tables not in the
    map render empty — fail-closed)

Controller becomes a thin caller again: drops EntityManager injection,
the two static lookup maps, the loadCurrentEntitySnapshot /
loadCurrentValuesFor / findFieldRecursive helpers, and the unused
value-holder imports (Analyzer, Dictionary, Panel, SiteInformation,
Test, TestSection, TypeOfSample) that were only there for the moved
REF_TABLE_TO_ENTITY_CLASS. Adds @Autowired AuditEntitySnapshotService.

No behavior change — same call shape, same allow-list, same fallback
semantics.
